Decided to remove datetimepicker
Could not make the date fiends validated with the bootstrap ui
datetimepicker i chose. For now replaced with simple inputs and will go
back to that soon. Also removed users to simplify

Note: the date looks off by a day when no time is set in the db date
field. Need to look into that!